# Conflicts:
#   back/src/main/java/com/web/curation/controller/RecellController.java
#   back/src/main/java/com/web/curation/model/Recell.java
#   front/src/views/DM.vue

+++ front/src/views/Settings.vue
@@ -3,7 +3,7 @@
     <h1 class="settings-name">계정설정</h1>
     <div class="settings-btn-area">
       <button class="btn change-password-btn settings-btn" @click="goPassword">비밀번호 변경</button>
-      <button class="btn change-search-btn settings-btn">검색내역 지우기</button>
+      <button class="btn change-search-btn settings-btn" @click="deleteSearch">검색내역 지우기</button>
       <button class="btn change-search-btn settings-btn" @click="deleteAllAlarm">알림내역 지우기</button>
       <button class="btn logout-btn settings-btn" @click="goLogout" id="customBtn">로그아웃</button>
       <button class="btn signout-btn settings-btn" @click="leave">회원 탈퇴</button>
@@ -15,6 +15,9 @@
 import "../components/css/settings.css"
 import { mapState, mapMutations, mapActions  } from 'vuex' 
 import axios from 'axios'
+import Swal from 'sweetalert2'
+import firebase from 'firebase'
+
 
 export default {
   name: 'Settings',
@@ -36,6 +39,25 @@ export default {
     ...mapMutations(['setToken', 'setLoggedIn', 'setUser']),
     ...mapActions([]),
     
+     deleteAtPath(path) {
+    var deleteList = [];
+    firebase.firestore().collection(path).get()
+    .then(snapshot => {
+      snapshot.forEach(doc=>{
+        console.log(doc.id,'=>',doc.data());
+        deleteList.push(doc.id);
+      });
+    })
+    .catch()
+    .finally(()=> {
+      console.log('delete=>',deleteList)
+      deleteList.forEach(docName =>{
+        console.log('docName=>',docName)
+        firebase.firestore().collection(path).doc(docName).delete();
+      })
+    })
+
+  },
     goLogout() {
       this.$cookies.remove('auth-token')
       this.$cookies.remove('auth-nickname')
@@ -89,6 +111,23 @@ export default {
     goPassword() {
       this.$router.push('/newpassword').catch(()=>{})
     },
+    deleteSearch() {
+      let tmpNick = this.$cookies.get('auth-nickname');
+      let uri = tmpNick;
+      let uri_enc = encodeURIComponent(uri);
+      let uri_dec = decodeURIComponent(uri_enc);
+      let resNick = uri_dec;
+
+      let nickname=resNick;
+      axios.delete(`https://i3b304.p.ssafy.io/api/search/deleteSearchHistory/${nickname}`
+      )
+      .then(Swal.fire(
+        '검색기록이 삭제되었습니다..',
+        '새로운 검색을 입력해주세요!',
+        'success'
+      ))
+      .catch()
+    },
     deleteAllAlarm(){
 
       let tmpNick = this.$cookies.get('auth-nickname');
@@ -97,14 +136,13 @@ export default {
       let uri_dec = decodeURIComponent(uri_enc);
       let resNick = uri_dec;
       let nickname=resNick;
-      console.log(nickname,1);
       const frm = new FormData();
       frm.append('recevier',nickname);
       axios.post('https://i3b304.p.ssafy.io/api/alarm/del',frm
       )
       .then(Swal.fire(
-        '알림기록이 삭제되었습니다..',
-        '새로운 알람을 기대하세요!',
+        '알람기록이 삭제되었습니다..',
+        '새로운 알람을 기대해주세요!',
         'success'
       ))
       .catch()
@@ -117,17 +155,38 @@ export default {
       let uri_enc = encodeURIComponent(uri);
       let uri_dec = decodeURIComponent(uri_enc);
       let resNick = uri_dec;
-
       let nickname=resNick;
+
+      
+      
       const formData=new FormData();
-      console.log(nickname);
       formData.append('nickname',nickname);
-      axios.delete('https://i3b304.p.ssafy.io/api/account/delete',{
+
+      Swal.fire({
+        title: '정말로 탈퇴하실거에요?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '네',
+        cancelButtonText: '아니오',
+      }).then((result) => {
+        if (result.value) {
+        axios.get('https://i3b304.p.ssafy.io/api/chat/allChatList',{
+        params:{
+        username: nickname,
+       }
+       })
+       .then((data)=>{
+        data.data.object.forEach(obj=>{
+          this.deleteAtPath(obj.roomname);
+        });
+       })
+       .catch()
+          axios.delete('https://i3b304.p.ssafy.io/api/account/delete',{
         data:formData,
-        
       })
       .then((data)=>{
-        console.log(data)
         if(data.data.data=="success"){
           ref.$cookies.remove('auth-token')
           ref.$cookies.remove('auth-nickname')
@@ -136,11 +195,19 @@ export default {
           ref.setUser(null)
           ref.$router.push('/').catch(()=>{})
         }
-        else{
-          console.log("fail..");
-        }
       })
       .catch()
+          Swal.fire(
+            '완료되었습니다.',
+          ).then((result2) => {
+            if (result2.value) {
+              this.$router.push('/').catch(()=>{})
+                }
+              })
+            }
+          }
+        )
+      
     },
   },
 }
